+++ blockstack-sdk/src/androidTest/java/org/blockstack/android/sdk/BlockstackSignInTest.kt
@@ -7,9 +7,10 @@ import kotlinx.coroutines.runBlocking
 import org.blockstack.android.sdk.model.BlockstackConfig
 import org.blockstack.android.sdk.model.toBlockstackConfig
 import org.blockstack.android.sdk.test.TestActivity
-import org.hamcrest.CoreMatchers
-import org.hamcrest.MatcherAssert
+import org.hamcrest.MatcherAssert.assertThat
 import org.hamcrest.Matchers
+import org.hamcrest.Matchers.`is`
+import org.hamcrest.Matchers.notNullValue
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
@@ -43,21 +44,23 @@ class BlockstackSignInTest {
     fun generateAndStoreTransitKeyReturnsTheCorrectKey() {
         val key = signIn.generateAndStoreTransitKey()
         val storedKey = sessionStore.sessionData.json.getString("transitKey")
-        MatcherAssert.assertThat(key, CoreMatchers.`is`(storedKey))
+        assertThat(key, `is`(storedKey))
+        val storedKey2 = sessionStore.getTransitPrivateKey()
+        assertThat(key, `is`(storedKey2))
     }
 
     @Test
     fun makeAuthRequestReturnsValidRequestToken() {
         val key = signIn.generateAndStoreTransitKey()
         val authRequest = runBlocking {
-             signIn.makeAuthRequest(key, Date(System.currentTimeMillis() + 3600000).time, mapOf(Pair("solicitGaiaHubUrl", true)))
+            signIn.makeAuthRequest(key, Date(System.currentTimeMillis() + 3600000).time, mapOf("solicitGaiaHubUrl" to true))
         }
-        MatcherAssert.assertThat(authRequest, Matchers.startsWith("ey"))
+        assertThat(authRequest, Matchers.startsWith("ey"))
 
         val token = Blockstack().decodeToken(authRequest)
-        val payload = token.second.optJSONObject("payload")
-        Log.d("payload", payload.toString())
-        MatcherAssert.assertThat(payload.optString("domain_name"), CoreMatchers.`is`(config.appDomain.toString()))
-        MatcherAssert.assertThat(payload.optBoolean("solicitGaiaHubUrl"), CoreMatchers.`is`(true))
+        val payload = token.second
+        assertThat(payload, `is`(notNullValue()))
+        assertThat(payload.optString("domain_name"), `is`(config.appDomain.toString()))
+        assertThat(payload.optBoolean("solicitGaiaHubUrl"), `is`(true))
     }
 }