+++ blockstack-sdk/src/main/java/org/blockstack/android/sdk/Defaults.kt
@@ -2,5 +2,5 @@ package org.blockstack.android.sdk
 
 const val BLOCKSTACK_DEFAULT_GAIA_HUB_URL = "https://hub.blockstack.org"
 const val DEFAULT_CORE_API_ENDPOINT = "https://core.blockstack.org"
-
+const val DEFAULT_BLOCKSTACK_ID_HOST = "https://browser.blockstack.org/auth"
 const val VERSION = "1.3.1"